Move the database configuration into the EmbeddedAnsible class
There's no reason for this to be in the console or elsewhere.
We can manage the creation of the database the same way we manage
the secret key. Create it if we think we need to.

We are using the presence of the database password to determine
if the database needs to be created.


(transferred from ManageIQ/manageiq@f9705124bf24831981a1a5540d4e4aca5984b6f6)