Implement a better worker lifecycle for the embedded_ansible_worker
Only heartbeat when the locally running ansible service is alive.

Also, start the service if it isn't alive and all the processes are
not running (EmbeddedAnsible.running?).

We are aware that this opens the potential for a race condition when
the worker is killed and the server starts a replacement before the
original worker is completely down.

The combination of implementing EmbeddedAnsibleWorker#kill as #stop
and starting the service every time we can't heartbeat should allow
the new worker to recover even if the stop is run after the new worker
runs EmbeddedAnsible.start

If this does become a problem we will need to implement some new
mechanisms for dealing with such "singleton" workers.


(transferred from ManageIQ/manageiq@678c4e731c8ddbdb71dfd3f6677112d314b2327a)