Use the new setup script argument types
A change was made to pass options that come to the setup script
after a "--" straight through, unaltered to the ansible-playbook
command.

This is how we will get access to the --tags and --skip-tags options
that we need.


(transferred from ManageIQ/manageiq@b12c8fd5a73f58e6bf0e0f8b4fbdb475ccb94eab)