Fix up the initial embedded ansible environment from the EmbeddedAnsibleWorker::Runner
This creates a concern around creating and removing objects
in the embedded ansible instance.

In particular we remove all of the demo data that is initialized
for us during setup and we create just the objects that we need
and save the ids to the provider.


(transferred from ManageIQ/manageiq@8322c6635b38ee5b4c9e3543b9b584d6f783823a)