Add a concern for storing and accessing default embedded ansible object ids
This will allow us to determine the ansible side object to relate
new objects to.

This also frees us from having to store an extra flag in our inventory
to identify which objects are the "system created" ones and which
are the other objects that may be artifacts of running jobs.

https://www.pivotaltracker.com/story/show/140098929


(transferred from ManageIQ/manageiq@592121989109bede0e7e30b110ef3d8193efdaf7)