Only run the setup playbook the first time we start
It needs to run once to put files in place and such, but
it restarts services in a way that could cause us to operate on
a running stack only to have it restart from under us.

This change makes us only run the setup playbook once so the
chance of hitting this kind of issue should be much smaller.

This effectively combines the .configure and .start methods
so that we detect when we are in the first configuration state
vs just starting up the services.

https://bugzilla.redhat.com/show_bug.cgi?id=1455063


(transferred from ManageIQ/manageiq@060a0c4998c32ac4a85cb5b593a4f278bd6cfc85)