Handle additional case for /etc/tower/SECRET_KEY
Previously if we had a value in the database, but the file
didn't exist on the filesystem .configured? would raise an error
when it should really just return false and we will write out
the value from the database to the filesystem.

https://bugzilla.redhat.com/show_bug.cgi?id=1439783
https://bugzilla.redhat.com/show_bug.cgi?id=1458886


(transferred from ManageIQ/manageiq@42eb2f8deefcf4b5390d7ac31dbaa195f289afcf)